feat: abstracted `Overview` component to shared; refactorings to `Infohub`
- Added announcement section, editable in `general` disposition
