+++ src/app/shared/_modules.tsx
@@ -3,3 +3,4 @@ export * from './Footer'
 export * from './Gallery'
 export * from './IntroFade'
 export * from './Navbar'
+export * from './InfoHub'
